Moved TestHelpers.ts to test root
