+++ dev/src/__tests__/e2e/E2ECourseVisibility.test.tsx
@@ -1,5 +1,5 @@
 import { By, until } from 'selenium-webdriver'
-import { setupDrivers } from '../testHelpers/TestHelpers'
+import { setupDrivers } from '../TestHelpers'
 
 
 describe("End to End course visibility", () => {
